requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
arcgis=1.9.0=py38_2172
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py38h7f8727e_0
attrs=21.4.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.4=py38h06a4308_0
bottleneck=1.3.2=py38heb32a55_1
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py38h27cfd23_1003
ca-certificates=2021.10.8=ha878542_0
certifi=2021.10.8=py38h578d9bd_1
cffi=1.15.0=py38hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=3.4.8=py38hd23ed53_0
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py38h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.3=py38_0
expat=2.4.4=h295c915_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
glib=2.69.1=h4ff587b_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.8.2=py38h06a4308_0
importlib_metadata=4.8.2=hd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.4.1=py38h06a4308_1
ipython=7.31.1=py38h06a4308_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py38h06a4308_1
jeepney=0.7.1=pyhd3eb1b0_0
jinja2=3.0.2=pyhd3eb1b0_0
jpeg=9d=h7f8727e_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter_client=6.1.7=py_1
jupyter_core=4.9.1=py38h06a4308_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
keyring=23.4.0=py38h06a4308_0
kiwisolver=1.3.2=py38h295c915_0
krb5=1.19.2=hac12032_0
ld_impl_linux-64=2.35.1=h7274673_9
lerc=3.0=pyh39e3cac_0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.3.0=h04863e7_1
libuuid=1.0.3=h7f8727e_2
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
lz4-c=1.9.3=h295c915_1
markupsafe=2.0.1=py38h27cfd23_0
matplotlib=3.5.0=py38h06a4308_0
matplotlib-base=3.5.0=py38h3ed280b_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py38h7b6447c_1000
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
munkres=1.1.4=py_0
nbconvert=5.6.1=py38_0
nbformat=5.1.3=py_1
ncurses=6.3=h7f8727e_2
notebook=6.1.4=py38_1
ntlm-auth=1.4.0=py_0
numexpr=2.8.1=py38h6abb31d_0
numpy=1.21.2=py38h20f2e39_0
numpy-base=1.21.2=py38h79a1101_0
oauthlib=3.2.0=pyhd3eb1b0_0
olefile=0.46=pyhd3eb1b0_0
openssl=1.1.1m=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.1=py38h295c915_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.3.2=py38_0
pip=21.2.4=py38h06a4308_0
prometheus_client=0.8.0=py_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pycparser=2.21=pyhd3eb1b0_0
pygments=2.9.0=py_0
pyjwt=2.1.0=py38h06a4308_0
pykerberos=1.2.1=py38h7f8727e_3
pyopenssl=19.1.0=py38_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py38h05f1152_4
pyrsistent=0.18.0=py38heee7806_0
pyshp=2.1.3=pyhd3eb1b0_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.12=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-decouple=3.6=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytz=2021.3=pyhd3eb1b0_0
pyzmq=22.3.0=py38h295c915_2
qt=5.9.7=h5867ecd_1
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd3eb1b0_0
requests-kerberos=0.12.0=py38_0
requests-oauthlib=1.3.0=py_0
requests-toolbelt=0.9.1=pyhd3eb1b0_0
requests_ntlm=1.1.0=py_0
secretstorage=3.3.1=py38h06a4308_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=58.0.4=py38h06a4308_0
sip=4.19.13=py38h295c915_0
six=1.15.0=py_0
sqlite=3.37.2=hc218d9a_0
terminado=0.13.1=py38h06a4308_0
testpath=0.5.0=pyhd3eb1b0_0
tk=8.6.11=h1ccaba5_0
tornado=6.1=py38h27cfd23_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
ujson=4.2.0=py38h295c915_0
urllib3=1.26.8=pyhd3eb1b0_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py38h06a4308_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zipp=3.7.0=pyhd3eb1b0_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
